Rename network_test.py to neural_network_test.py
Tests the neural network from homework_4_NN, the network implemented does not train or function as well as the Sklearn implementation. The code is modified from an updated implementation of an online ML textbook (citation will be added when I find it). The main modification from the previous code was implementing softmax on the output layer.